refactor: removed re-exports of internal classes
