+++ tests/safeds/data/image/containers/test_image_list.py
@@ -5,7 +5,10 @@ import tempfile
 import pytest
 import torch
 from safeds.config import _get_device
-from safeds.data.image.containers import Image, ImageList, _EmptyImageList, _MultiSizeImageList, _SingleSizeImageList
+from safeds.data.image.containers import Image, ImageList
+from safeds.data.image.containers._empty_image_list import _EmptyImageList
+from safeds.data.image.containers._multi_size_image_list import _MultiSizeImageList
+from safeds.data.image.containers._single_size_image_list import _SingleSizeImageList
 from safeds.data.tabular.containers import Table
 from safeds.exceptions import DuplicateIndexError, IllegalFormatError, IndexOutOfBoundsError, OutOfBoundsError
 from syrupy import SnapshotAssertion
